Prevent event bubbling
If you change the action to an `onclick={{action...` then the act of
clicking a button will trigger the form submission due to event
bubbling. This return false prevents that.

This would be needed if we find that an onclick or touchend is better
then the ember sugar we use now.