feat(onUpdate): Separate onSubmit and onUpdate actions
Prior to this change, we had an onSubmit action and an autoSave flag
which would debounce the call to onSubmit when the use updates a field.

This change separates the two so that onUpdate is debounce when the user
changes things and the onSubmit only happens when the user actually
submits the form. That is of course as long as validations pass.