feat(parser): Add an upgraded Schema Form Parser
Prior to this change, there was a fundamental misunderstanding of the
way the forms portion of the definition is meant to be used. Because of
this we were unable to support deterministic ordering and grouping.

This change is a complete rewrite of the schema parsing which will
support the correct format while keeping backwards compatibility
(deprecated).

https://github.com/json-schema-form/json-schema-form/wiki/Documentation