refactor(componentName): Move componentName into SchemaFormParser
Prior to this change, the componentName was being grafted onto the
results of the parser to support presentation needs. Since the idea
behind the parser was to provide a sanitized data format useable by
presentation logic this separation seemed unnecessary.

This change moves the grafting into the parser but keeps the component
lookup as a callback.